+++ plone/app/users/browser/passwordpanel.py
@@ -1,13 +1,13 @@
+# -*- coding: utf-8 -*-
 from Acquisition import aq_inner
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import PloneMessageFactory as _
 from Products.statusmessages.interfaces import IStatusMessage
 from plone.app.users.browser.account import AccountPanelForm
+from plone.app.users.utils import notifyWidgetActionExecutionError
 from z3c.form import button
 from zope import schema
 from zope.interface import Interface
-
-from ..utils import notifyWidgetActionExecutionError
 
 
 class IPasswordSchema(Interface):
@@ -18,7 +18,7 @@ class IPasswordSchema(Interface):
         description=_(
             u'help_current_password',
             default=u'Enter your current password.'),
-        #constraint=checkCurrentPassword,
+        # constraint=checkCurrentPassword,
     )
 
     new_password = schema.Password(