+++ plone/app/users/browser/register.py
@@ -1,8 +1,10 @@
+# -*- coding: utf-8 -*-
 from AccessControl import getSecurityManager
 from Products.CMFCore.interfaces import ISiteRoot
 from Products.CMFCore.permissions import ManagePortal
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import PloneMessageFactory as _
+from Products.CMFPlone.interfaces import ISecuritySchema
 from Products.CMFPlone.utils import normalizeString
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.statusmessages.interfaces import IStatusMessage
@@ -10,17 +12,23 @@ from ZODB.POSException import ConflictError
 from plone.app.layout.navigation.interfaces import INavigationRoot
 from plone.app.users.browser.interfaces import ILoginNameGenerator
 from plone.app.users.browser.interfaces import IUserIdGenerator
+from plone.app.users.schema import IAddUserSchema
+from plone.app.users.schema import ICombinedRegisterSchema
+from plone.app.users.schema import IRegisterSchema
+from plone.app.users.utils import notifyWidgetActionExecutionError
 from plone.app.users.utils import uuid_userid_generator
 from plone.autoform.form import AutoExtensibleForm
 from plone.autoform.interfaces import OMITTED_KEY
 from plone.autoform.interfaces import ORDER_KEY
 from plone.protect import CheckAuthenticator
+from plone.registry.interfaces import IRegistry
 from z3c.form import button
 from z3c.form import field
 from z3c.form import form
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from z3c.form.interfaces import DISPLAY_MODE
 from zExceptions import Forbidden
+from zope.component import getAdapter
 from zope.component import getMultiAdapter
 from zope.component import (
     getUtility,
@@ -63,6 +71,12 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         # adapter for it
         provideAdapter(RegisterSchemaProvider, (INavigationRoot,), self.schema)
 
+    def _get_security_settings(self):
+        """Return security settings from the registry."""
+        registry = getUtility(IRegistry)
+        return registry.forInterface(ISecuritySchema, prefix="plone")
+
+
     def render(self):
         if self._finishedRegister:
             return self.context.unrestrictedTraverse('registered')()
@@ -75,26 +89,27 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         """
         portal_props = getToolByName(self.context, 'portal_properties')
         props = portal_props.site_properties
-        use_email_as_login = props.getProperty('use_email_as_login')
+        settings = self._get_security_settings()
+        use_email_as_login = settings.use_email_as_login
 
         # Ensure all listed fields are in the schema
         registration_fields = [f for f in props.getProperty(
             'user_registration_fields', []) if f in self.schema]
 
         # Check on required join fields
-        if not 'username' in registration_fields and not use_email_as_login:
+        if 'username' not in registration_fields and not use_email_as_login:
             registration_fields.insert(0, 'username')
 
         if 'username' in registration_fields and use_email_as_login:
             registration_fields.remove('username')
 
-        if not 'email' in registration_fields:
+        if 'email' not in registration_fields:
             # Perhaps only when use_email_as_login is true, but also
             # for some other cases; the email field has always been
             # required.
             registration_fields.append('email')
 
-        if not 'password' in registration_fields:
+        if 'password' not in registration_fields:
             if 'username' in registration_fields:
                 base = registration_fields.index('username')
             else:
@@ -102,12 +117,12 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
             registration_fields.insert(base + 1, 'password')
 
         # Add password_ctl after password
-        if not 'password_ctl' in registration_fields:
+        if 'password_ctl' not in registration_fields:
             registration_fields.insert(
                 registration_fields.index('password') + 1, 'password_ctl')
 
         # Add email_me after password_ctl
-        if not 'mail_me' in registration_fields:
+        if 'mail_me' not in registration_fields:
             registration_fields.insert(
                 registration_fields.index('password_ctl') + 1, 'mail_me')
 
@@ -171,7 +186,7 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         1. We query a utility, so integrators can register a hook to
            generate a user id using their own logic.
 
-        2. If use_uuid_as_userid is set in the site_properties, we
+        2. If use_uuid_as_userid is set in the registry, we
            generate a uuid.
 
         3. If a username is given and we do not use email as login,
@@ -210,9 +225,8 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
                 data['user_id'] = userid
                 return userid
 
-        portal_props = getToolByName(self.context, 'portal_properties')
-        props = portal_props.site_properties
-        if props.getProperty('use_uuid_as_userid'):
+        settings = self._get_security_settings()
+        if settings.use_uuid_as_userid:
             userid = uuid_userid_generator()
             data['user_id'] = userid
             return userid
@@ -221,13 +235,15 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         userid = data.get('username')
         if userid:
             # If we are not using email as login, then this user name is fine.
-            if not props.getProperty('use_email_as_login'):
+            if not settings.use_email_as_login:
                 data['user_id'] = userid
                 return userid
 
         # First get a default value that we can return if we cannot
         # find anything better.
-        default = data.get('username') or data.get('email') or ''
+        pas = getToolByName(self.context, 'acl_users')
+        email = pas.applyTransform(data.get('email'))
+        default = data.get('username') or email or ''
         data['user_id'] = default
         fullname = data.get('fullname')
         if not fullname:
@@ -297,11 +313,9 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         login_name = data.get('username')
         login_name = pas.applyTransform(login_name)
         data['login_name'] = login_name
-        portal_props = getToolByName(self.context, 'portal_properties')
-        props = portal_props.site_properties
-        use_email_as_login = props.getProperty('use_email_as_login')
+        settings = self._get_security_settings()
         # If we are not using email as login, then this user name is fine.
-        if not use_email_as_login:
+        if not settings.use_email_as_login:
             return login_name
 
         # We use email as login.
@@ -331,9 +345,6 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         form_field_names = [f for f in self.fields]
 
         portal = getUtility(ISiteRoot)
-        portal_props = getToolByName(self.context, 'portal_properties')
-        props = portal_props.site_properties
-        use_email_as_login = props.getProperty('use_email_as_login')
 
         # passwords should match
         if 'password' in form_field_names:
@@ -352,7 +363,7 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         # Password field checked against RegistrationTool
         if 'password' in form_field_names:
             # Skip this check if password fields already have an error
-            if not 'password' in error_keys:
+            if 'password' not in error_keys:
                 password = data.get('password')
                 if password:
                     # Use PAS to test validity
@@ -361,7 +372,8 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
                         notifyWidgetActionExecutionError(action,
                                                          'password', err_str)
 
-        if use_email_as_login:
+        settings = self._get_security_settings()
+        if settings.use_email_as_login:
             username_field = 'email'
         else:
             username_field = 'username'
@@ -386,7 +398,7 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
 
         # check if username is valid
         # Skip this check if username was already in error list
-        if not username_field in error_keys:
+        if username_field not in error_keys:
             # user id may not be the same as the portal id.
             if user_id == portal.getId():
                 err_str = _(u"This username is reserved. Please choose a "
@@ -395,14 +407,14 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
                                                  username_field, err_str)
 
         # Check if user id is allowed by the member id pattern.
-        if not username_field in error_keys:
+        if username_field not in error_keys:
             if not registration.isMemberIdAllowed(user_id):
                 err_str = _(u"The login name you selected is already in use "
                             "or is not valid. Please choose another.")
                 notifyWidgetActionExecutionError(action,
                                                  username_field, err_str)
 
-        if not username_field in error_keys:
+        if username_field not in error_keys:
             # Check the uniqueness of the login name, not only when
             # use_email_as_login is true, but always.
             pas = getToolByName(self, 'acl_users')
@@ -413,7 +425,7 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
                 notifyWidgetActionExecutionError(action,
                                                  username_field, err_str)
 
-        if 'password' in form_field_names and not 'password' in error_keys:
+        if 'password' in form_field_names and 'password' not in error_keys:
             # Admin can either set a password or mail the user (or both).
             if not (data['password'] or data['mail_me']):
                 err_str = _('msg_no_password_no_mail_me',
@@ -498,7 +510,8 @@ class BaseRegistrationForm(AutoExtensibleForm, form.Form):
         # set member properties
         self.applyProperties(user_id, data)
 
-        if data.get('mail_me') or (portal.validate_email and
+        settings = self._get_security_settings()
+        if data.get('mail_me') or (not settings.enable_user_pwd_choice and
                                    not data.get('password')):
             # We want to validate the email address (users cannot
             # select their own passwords on the register form) or the
@@ -608,9 +621,11 @@ class RegistrationForm(BaseRegistrationForm):
         ctrlOverview = getMultiAdapter((portal, self.request),
                                        name='overview-controlpanel')
 
-        # hide form iff mailhost_warning == True and validate_email == True
+        settings = self._get_security_settings()
+        # hide form if mailhost_warning == True and
+        # enable_user_pwd_choice == False
         return not (ctrlOverview.mailhost_warning() and
-                    portal.getProperty('validate_email', True))
+                    not settings.enable_user_pwd_choice)
 
     def updateFields(self):
         if not self.showForm:
@@ -622,8 +637,8 @@ class RegistrationForm(BaseRegistrationForm):
         defaultFields = field.Fields(self.fields)
 
         # Can the user actually set his/her own password?
-        portal = getUtility(ISiteRoot)
-        if portal.getProperty('validate_email', True):
+        settings = self._get_security_settings()
+        if not settings.enable_user_pwd_choice:
             # No? Remove the password fields.
             defaultFields = defaultFields.omit('password', 'password_ctl')
         else:
@@ -642,8 +657,8 @@ class RegistrationForm(BaseRegistrationForm):
             return
 
         super(RegistrationForm, self).updateWidgets()
-        portal = getUtility(ISiteRoot)
-        if portal.getProperty('validate_email', True):
+        settings = self._get_security_settings()
+        if not settings.enable_user_pwd_choice:
             # Show a message indicating that a password reset link
             # will be mailed to the user.
             widget = self.widgets['mail_me']
@@ -669,7 +684,7 @@ class AddUserForm(BaseRegistrationForm):
         defaultFields = field.Fields(self.fields)
 
         # The mail_me field needs special handling depending on the
-        # validate_email property and on the correctness of the mail
+        # enable_user_pwd_choice setting and on the correctness of the mail
         # settings.
         portal = getUtility(ISiteRoot)
         ctrlOverview = getMultiAdapter((portal, self.request),
@@ -681,7 +696,8 @@ class AddUserForm(BaseRegistrationForm):
             # will check that at least one of the options is chosen.
             defaultFields['password'].field.required = False
             defaultFields['password_ctl'].field.required = False
-            if portal.getProperty('validate_email', True):
+            settings = self._get_security_settings()
+            if not settings.enable_user_pwd_choice:
                 defaultFields['mail_me'].field.default = True
             else:
                 defaultFields['mail_me'].field.default = False