+++ plone/app/users/tests/__init__.py
@@ -1,2 +1,3 @@
+# -*- coding: utf-8 -*-
 # Other packages may find this useful
 from plone.app.users.tests.base import BaseTestCase