+++ plone/app/users/tests/test_member_search.py
@@ -1,3 +1,6 @@
+# -*- coding: utf-8 -*-
+from plone.app.users.browser.membersearch import extractCriteriaFromRequest
+
 import unittest
 
 
@@ -11,7 +14,6 @@ class TestMemberSearch(unittest.TestCase):
             'form.widgets.roles-empty-marker': True,
         }
 
-        from plone.app.users.browser.membersearch import extractCriteriaFromRequest
         result = extractCriteriaFromRequest(data)
 
-        self.assertEqual(result, {"something": u'any form value'})
+        self.assertEqual(result, {'something': u'any form value'})