+++ plone/app/users/utils.py
@@ -1,11 +1,11 @@
-from zope.component import getUtility
-import zope.event
-from zope.interface import Invalid
-
+# -*- coding: utf-8 -*-
+from plone.uuid.interfaces import IUUIDGenerator
 from z3c.form.action import ActionErrorOccurred
 from z3c.form.interfaces import WidgetActionExecutionError
+from zope.component import getUtility
+from zope.interface import Invalid
 
-from plone.uuid.interfaces import IUUIDGenerator
+import zope.event
 
 
 def uuid_userid_generator(data=None):
@@ -18,4 +18,7 @@ def uuid_userid_generator(data=None):
 def notifyWidgetActionExecutionError(action, widget, err_str):
     zope.event.notify(
         ActionErrorOccurred(
-            action, WidgetActionExecutionError(widget, Invalid(err_str))))
+            action,
+            WidgetActionExecutionError(widget, Invalid(err_str))
+        )
+    )