+++ plone/app/users/vocabularies.py
@@ -1,10 +1,16 @@
+# -*- coding: utf-8 -*-
 from AccessControl import getSecurityManager
-
+from Products.CMFCore.permissions import ManagePortal
+from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.utils import normalizeString
+from Products.CMFPlone.utils import safe_unicode
+from plone.app.users.browser.register import RegistrationForm
 from zope.interface import implements
 from zope.component import getUtility
 from zope.schema import getFieldNames
 from zope.schema.interfaces import IVocabularyFactory
-from zope.schema.vocabulary import SimpleVocabulary, SimpleTerm
+from zope.schema.vocabulary import SimpleVocabulary
+from zope.schema.vocabulary import SimpleTerm
 from zope.site.hooks import getSite
 
 from Products.CMFCore.utils import getToolByName